requirements.txt
customtkinter
Pillow
keyboard
